+++ src/components/GalleryCarouselSlider/GalleryCarouselSlider.css
@@ -3,3 +3,33 @@
 .sliderWrapper {
   height: 100%;
 }
+
+.iconArrow {
+  width: 60px;
+  height: 60px;
+  stroke-width: 0px;
+  stroke: var(--matterColorLight);
+  fill: var(--matterColorLight);
+}
+
+.carouselArrow {
+  position: absolute;
+  display: flex;
+  align-items: center;
+  height: 100%;
+  width: 20%;
+  top: 0;
+  border: none;
+  background: rgb(255, 255, 255);
+  outline: none;
+  cursor: pointer;
+}
+.nextSlide {
+  right: 0;
+  background: linear-gradient(90deg, rgba(255, 255, 255, 0) 0%, rgba(0, 0, 0, 0.5) 100%);
+}
+.prevSlide {
+  justify-content: flex-end;
+  left: 0;
+  background: linear-gradient(-90deg, rgba(255, 255, 255, 0) 0%, rgba(0, 0, 0, 0.5) 100%);
+}
